Added accurate distance volume curve functions
Direct translation of the C++ code that I had, into python. A lot of things could be written better.
First commit is just so I can get eyes on the code and see what everyone thinks.

I do need to rename some variables and reduce the overall comments.
But I am waiting to see if there is any feedback on what is here. Maybe some people can improve on what is here.